requirements.txt
crewai==0.22.5
crewai[tools]
langchain_openai
python-dotenv
langchain-groq